added enums, updated error handling, new markdown, added account status

+++ GoBankDB/src/main/java/com/chadwick/GoBankDB/service/AccountService.java
@@ -3,15 +3,17 @@ package com.chadwick.GoBankDB.service;
 import com.chadwick.GoBankDB.dto.AccountDTO;
 import com.chadwick.GoBankDB.entity.Account;
 import com.chadwick.GoBankDB.entity.Customer;
+import com.chadwick.GoBankDB.enums.AccountStatus;
+import com.chadwick.GoBankDB.exception.InternalServerException;
+import com.chadwick.GoBankDB.exception.NotFoundException;
 import com.chadwick.GoBankDB.model.Name;
 import com.chadwick.GoBankDB.repository.AccountRepository;
 import com.chadwick.GoBankDB.repository.CustomerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import org.springframework.web.server.ResponseStatusException;
-
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @Service
 public class AccountService {
@@ -24,36 +26,71 @@ public class AccountService {
         return accountRepository.findAll();
     }
 
-    public Account getAccountByID(long accountId){
-        return accountRepository.findById(accountId).get();
+    public AccountDTO getAccountByID(long accountId){
+        try{
+            if (!accountRepository.existsById(accountId)){
+                throw new NotFoundException("Account " + accountId + " does not exist.");
+            }
+            Account acc = accountRepository.findById(accountId).get();
+            return new AccountDTO(
+                    acc.getAccountId(),
+                    acc.getAccountOwnerId(),
+                    acc.getAccountOwnerName(),
+                    acc.getBalance(),
+                    acc.getAccountStatus(),
+                    acc.getTransactionIDs()
+            );
+        } catch (Exception e){
+            if(e instanceof NotFoundException){
+                throw e;
+            }
+            throw new InternalServerException(e.getMessage());
+        }
     }
 
     public List<Account> getAccountsByOwnerId(int accountOwnerId){
-        return accountRepository.findAccountsByOwnerId(accountOwnerId);
+        try{
+            return accountRepository.findAccountsByOwnerId(accountOwnerId);
+        } catch (Exception e){
+            throw new InternalServerException(e.getMessage());
+        }
     }
 
     public AccountDTO createAccount(Account account, int accountOwnerID){
-        // when creating a new account, the customer ID is taken in to add to the account object and is used to query the customer to then add the account id to the list in the customer entity
-        account.setAccountOwnerId(accountOwnerID);
-        Account acc = accountRepository.save(account);
-        Customer customer = customerRepository.findById(accountOwnerID).get();
-        List<Long> accountIDs = customer.getAccountIDs();
-        accountIDs.add(acc.getAccountId());
-        customer.setAccountIDs(accountIDs);
-        customerRepository.save(customer);
-        return new AccountDTO(
-                acc.getAccountId(),
-                acc.getAccountOwnerId(),
-                acc.getAccountOwnerName(),
-                acc.getBalance(),
-                acc.getTransactionIDs()
-        );
+        try{
+            // when creating a new account, the customer ID is taken in to add to the account object and is used to query the customer to then add the account id to the list in the customer entity
+            account.setAccountOwnerId(accountOwnerID);
+            Account acc = accountRepository.save(account);
+            if (!customerRepository.existsById(accountOwnerID)){
+                throw new NotFoundException("Account owner ID " + accountOwnerID + " not found");
+            }
+            Customer customer = customerRepository.findById(accountOwnerID).get();
+            List<Long> accountIDs = customer.getAccountIDs();
+            accountIDs.add(acc.getAccountId());
+            customer.setAccountIDs(accountIDs);
+            customerRepository.save(customer);
+            return new AccountDTO(
+                    acc.getAccountId(),
+                    acc.getAccountOwnerId(),
+                    acc.getAccountOwnerName(),
+                    acc.getBalance(),
+                    acc.getAccountStatus(),
+                    acc.getTransactionIDs()
+            );
+        } catch (Exception e) {
+            if(e instanceof NotFoundException){
+                throw e;
+            }
+            throw new InternalServerException(e.getMessage());
+        }
     }
 
     public Account updateAccount(long id, Account updates){
         try{
+            if (!accountRepository.existsById(id)){
+                throw new NotFoundException("Account " + id + " does not exist.");
+            }
             Account account = accountRepository.findById(id).get();
-
             if(updates.getAccountOwnerName() != null){
                 Name update = updates.getAccountOwnerName();
                 Name name = account.getAccountOwnerName();
@@ -76,18 +113,45 @@ public class AccountService {
             if(updates.getBalance() != null){
                 account.setBalance(updates.getBalance());
             }
+            if(updates.getAccountStatus() != null){
+                account.setAccountStatus(updates.getAccountStatus());
+                //check if the account status is now CLOSED, if so then account ID is removed from Customer accounts list
+                if(updates.getAccountStatus() == AccountStatus.CLOSED){
+                    Customer customer = customerRepository.findById(account.getAccountOwnerId()).get();
+                    List<Long> accountIDs = customer.getAccountIDs();
+                    accountIDs.remove(account.getAccountId());
+                    customer.setAccountIDs(accountIDs);
+                    customerRepository.save(customer);
+                }
+            }
             if(updates.getTransactionIDs() != null){
                 account.setTransactionIDs(updates.getTransactionIDs());
             }
 
             return accountRepository.save(account);
         }catch (Exception e){
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, null, e);
+            if(e instanceof NotFoundException){
+                throw e;
+            }
+            throw new InternalServerException(e.getMessage());
         }
     }
 
-    public HttpStatus deleteAccount(long accountId){
-        accountRepository.deleteById(accountId);
-        return HttpStatus.OK;
+    public Map<String,Long> deleteAccount(long accountId){
+        try{
+            Map<String,Long> map = new HashMap<>();
+            if (!accountRepository.existsById(accountId)){
+                throw new NotFoundException("Account ID provided does not exist.");
+            }
+            accountRepository.deleteById(accountId);
+            map.put("accountId", accountId);
+            return map;
+
+        }catch (Exception e){
+            if(e instanceof NotFoundException){
+                throw e;
+            }
+            throw new InternalServerException(e.getMessage());
+        }
     }
 }
